mgp2csv: add heat info parsing
Public pages didn't have heat information. Add the ability to login
using curl and saved cookies, and get to the logged-in view of the race,
and parse pilors by heat, and group them.

Also add some empty rows at the bottom as scribble space.

Authentication is done using an 'auth' file in the current working
directory. See auth.sample on how to fill it in.